ai-logic/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
dill==0.3.7
fastapi==0.105.0
filelock==3.13.1
fonttools==4.47.2
fsspec==2023.12.2
gevent==23.9.1
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
hub-sdk==0.0.3
idna==3.6
imageio==2.33.1
imutils==0.5.4
iniconfig==2.0.0
Jinja2==3.1.3
kiwisolver==1.4.5
lazy_loader==0.3
MarkupSafe==2.1.4
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
opencv-python==4.8.1.78
packaging==23.2
pandas==2.2.0
Pillow==10.1.0
pluggy==1.4.0
psutil==5.9.8
py-cpuinfo==9.0.0
pydantic==2.5.3
pydantic_core==2.14.6
pyparsing==3.1.1
pytesseract==0.3.10
pytest==7.4.4
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.22.0
scipy==1.12.0
seaborn==0.13.1
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
thop==0.1.1.post2209072238
tifffile==2023.12.9
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
ultralytics==8.0.239
urllib3==2.1.0
uvicorn==0.27.0
zope.event==5.0
zope.interface==6.1
